pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agente-pesquisa"
version = "0.1.0"
description = "Monitor de campanhas financeiras com OpenClaw + Codex"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "markdown>=3.5.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
  "python-dateutil>=2.9.0",
  "jsonschema>=4.22.0",
  "requests>=2.32.0",
  "playwright>=1.52.0",
  "Pillow>=10.4.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
